Refactored how projects add directories to each other, added some fixes for compile errors in overlap tracker

+++ sector_grid_data_structure/sector_grid_index.h
@@ -1,11 +1,11 @@
 #pragma once
 #include "sector_grid_dimensions.h"
-#include "int_type_selection.h"
+#include "misc_utilities/int_type_selection.h"
 #include <concepts>
 #include <bit>
 #include <bitset>
 #include <limits>
-#include "vector_types.h"
+#include "vector_2d_math_utils/vector_types.h"
 #include "sector_grid_base_type_definition.h"
 
 namespace SectorGrid
@@ -18,7 +18,7 @@ namespace SectorGrid
 		typedef  MiscUtilities::uint_s<TSectorGridDimensions::sector_tile_count>::int_type_t tile_t; //type needed to hold the max value for tiles
 		typedef  MiscUtilities::uint_s<TSectorGridDimensions::sector_grid_count_max>::int_type_t sector_t; //type needed to hold max value for sectors 
 
-		static constexpr uint64 max_possible_tile_index = std::numeric_limits<tile_t>::max() * std::numeric_limits<sector_t>::max();
+		static constexpr size_t max_possible_tile_index = std::numeric_limits<tile_t>::max() * std::numeric_limits<sector_t>::max();
 
 		typedef  MiscUtilities::uint_s<max_possible_tile_index>::int_type_t combined_index; //type needed to hold combined types
